refactor(prd-1): address CodeRabbit nitpick comments for PR #12
Applied 7 nitpick improvements to enhance code quality and maintainability:

- Centralize CATEGORY_PAGES config into shared module (src/config/category-pages.js)
- Remove dead code: calculateCategoryActivity from sync-content.js
- Make spreadsheet config environment-driven (SPREADSHEET_ID, SHEET_NAME, RANGE)
- Add ISO date format support (YYYY-MM-DD) to parseDateToISO in both scripts
- Add fail-fast validation for required secrets in update-page-visibility.js
- Update PRD documentation to fix checkbox inconsistencies in Steps 6.2, 6.3, and Remaining Work section

Code improvements enhance flexibility (env-driven config), robustness (ISO date support),
error handling (fail-fast validation), and maintainability (DRY principle for config).
